Marina branch

+++ src/main/java/com/checkers/server/controller/UserRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 /**
  * Interface for User CRUD operations & queries
  */
-public interface UserRepository extends CrudRepository<User, Long> {
-	User findById(long id);
+public interface UserRepository extends CrudRepository<User, Integer> {
+	User findById(int id);
 
 	List<User> findByUsername(String userName);
